Remove pullRequestNum from desktop payload (#16)
It seems the Circle v2 API and/or the Circle Config v2.1 environment has problems resolving integer parameters. Have tried setting the parameter's `type` field in wp-desktop config.yml to both integer and string with no success:

> Type error for argument pullRequestNum: expected type: string, actual value: "39001" (type integer)

> Type error for argument pullRequestNum: expected type: integer, actual value: "39001" (type string)

This parameter is not sent back to the canary bridge by wp-desktop, so let's exclude it for now.

The build job will error out if a parameter in the sender's payload is not defined in the receiver's config.yml file, so we need to remove it from both the sending and receiving code. See corresponding change in github/Automattic/wp-desktop#765.